Is flags: break out server runnable unit-tests
