+++ code/ext/ext.lib.automerge/src/common/u.Is.TEST.ts
@@ -3,9 +3,8 @@ import { BroadcastChannelNetworkAdapter } from '@automerge/automerge-repo-networ
 import { IndexedDBStorageAdapter } from '@automerge/automerge-repo-storage-indexeddb';
 
 import { Is } from '.';
-import { Store } from '../Store';
 import { WebStore } from '../Store.Web';
-import { Doc, Test, TestDb, expect, type t } from '../test.ui';
+import { Test, TestDb, expect } from '../test.ui';
 import { RepoList } from '../ui/ui.RepoList';
 
 export default Test.describe('Is', (e) => {
@@ -20,54 +19,13 @@ export default Test.describe('Is', (e) => {
     storage: new IndexedDBStorageAdapter(TestDb.Unit.name),
   });
 
-  e.it('Is.automergeUrl', (e) => {
-    const store = Store.init();
-    const doc = store.repo.create();
-    expect(Is.automergeUrl(doc.url)).to.eql(true);
-    NON_OBJECTS.forEach((v) => expect(Is.automergeUrl(v)).to.eql(false));
-  });
-
-  e.it('Is.docRef', async (e) => {
-    type T = { count: number };
-    NON_OBJECTS.forEach((v) => expect(Is.docRef(v)).to.eql(false));
-    const store = Store.init();
-    const doc = await store.doc.getOrCreate<T>((d) => (d.count = 0));
-    expect(Is.docRef(doc)).to.eql(true);
-    store.dispose();
-  });
-
-  e.it('Is.store', (e) => {
-    const store = Store.init();
-    expect(Is.store(store)).to.eql(true);
-    NON_OBJECTS.forEach((value) => expect(Is.store(value)).to.eql(false));
-    store.dispose();
-  });
-
-  e.it('Is.storeIndex', async (e) => {
-    const store = Store.init();
-    const index = await Store.Index.init(store);
-    NON_OBJECTS.forEach((value) => expect(Is.storeIndex(value)).to.eql(false));
-    expect(Is.storeIndex(index)).to.eql(true);
-    store.dispose();
-  });
-
-  e.it('Is.webStore', (e) => {
-    expect(Is.webStore(WebStore.init({ storage: false }))).to.eql(true);
-    expect(Is.webStore(Store.init())).to.eql(false);
-    NON_OBJECTS.forEach((value) => expect(Is.store(value)).to.eql(false));
-  });
+  e.it('NOTE: other "Is" flag tests run in server-side unit-test file', (e) => {});
 
   e.it('Is.repo', (e) => {
     NON_OBJECTS.forEach((v) => expect(Is.repo(v)).to.eql(false));
     expect(Is.repo(repo1)).to.eql(true);
     expect(Is.repo(repo2)).to.eql(true);
     expect(Is.repo(repo3)).to.eql(true);
-  });
-
-  e.it('Is.repoIndex', (e) => {
-    const index: t.StoreIndex = { ['.meta']: Store.Index.meta, docs: [] };
-    expect(Is.repoIndex(index)).to.eql(true);
-    NON_OBJECTS.forEach((v) => expect(Is.repoIndex(v)).to.eql(false));
   });
 
   e.it('Is.repoListState', async (e) => {
@@ -99,21 +57,5 @@ export default Test.describe('Is', (e) => {
     expect(Is.storageSubsystem(repo1.storageSubsystem)).to.eql(false);
     expect(Is.storageSubsystem(repo2.storageSubsystem)).to.eql(false);
     expect(Is.storageSubsystem(repo3.storageSubsystem)).to.eql(true);
-  });
-
-  e.it('Is.broadcastChannel (network adapter)', (e) => {
-    const adapter = new BroadcastChannelNetworkAdapter();
-    expect(Is.broadcastChannel(adapter)).to.eql(true);
-    NON_OBJECTS.forEach((v) => expect(Is.broadcastChannel(v)).to.eql(false));
-  });
-
-  e.it('Is.namespace', async (e) => {
-    type TRoot = { count: number };
-    const store = Store.init();
-    const doc = await store.doc.getOrCreate<TRoot>((d) => (d.count = 0));
-    const ns = Doc.namespace(doc);
-    expect(Is.namespace(ns)).to.eql(true);
-    NON_OBJECTS.forEach((value) => expect(Is.namespace(value)).to.eql(false));
-    store.dispose();
   });
 });
